// api/model/userAdminUpdate.ts
/**
 * Generated by orval v8.4.1 🍺
 * Do not edit manually.
 * AI Shop Helper
 * OpenAPI spec version: 0.1.0
 */

/**
 * Schema for superuser updating any user.
 */
export interface UserAdminUpdate {
  email?: string | null;
  name?: string | null;
  password?: string | null;
  is_active?: boolean | null;
  is_superuser?: boolean | null;
}